+++ lib/packaging/artifactory.rb
@@ -408,15 +408,17 @@ module Pkg
     # Using the manifest provided by enterprise-dist, grab the appropropriate packages from artifactory based on md5sum
     # @param staging_directory [String] location to download packages to
     # @param manifest [File] JSON file containing information about what packages to download and the corresponding md5sums
-    def download_packages(staging_directory, manifest)
+    # @param path [String] path matching for where the package should be downloaded from. Defaults to '' (no custom path for searching)
+    def download_packages(staging_directory, manifest, path = '')
       check_authorization
       manifest.each do |dist, packages|
         puts "Grabbing the #{dist} packages from artifactory"
         packages.each do |name, info|
-          artifact_to_download = Artifactory::Resource::Artifact.checksum_search(md5: "#{info["md5"]}", repos: ["rpm_enterprise__local", "debian_enterprise__local"]).first
+          artifacts = Artifactory::Resource::Artifact.checksum_search(md5: "#{info["md5"]}", repos: ["rpm_enterprise__local", "debian_enterprise__local"])
+          artifact_to_download = artifacts.select { |artifact| artifact.download_uri.include? path }.first
           if artifact_to_download.nil?
             filename = info["filename"] || info["name"]
-            raise "Error: what the hell, could not find package #{filename} with md5sum #{info["md5"]}"
+            raise "Error: what the hell, could not find package #{filename} with md5sum #{info["md5"]} in #{path}"
           else
             full_staging_path = "#{staging_directory}/#{dist}"
             filename = info['filename'] || File.basename(artifact_to_download.download_uri)